+++ example/lib/example5_themes_playground/theme/flex_theme_dark.dart
@@ -208,6 +208,8 @@ FlexColorScheme flexColorSchemeDark(ThemeController controller, Color source) {
             // AppBar settings.
             appBarBackgroundSchemeColor:
                 controller.appBarBackgroundSchemeColorDark,
+            appBarScrolledUnderElevation:
+                controller.appBarScrolledUnderElevationDark,
             // TabBar settings.
             tabBarItemSchemeColor: controller.tabBarItemSchemeColorDark,
             tabBarIndicatorSchemeColor: controller.tabBarIndicatorDark,